Fixes maturin builds
Newer versions of `maturin` are more strict about what goes in Cargo.toml and pyproject.toml.

We need those to work so we can build wheels for other architectures than x86_64 (which is the whole point of this fork).

Tox is now working too. That invovled downgrading `setuptools` to work with `celery` 4.